VaultDetailQRCodeViewModel refactored
